requirements.txt
anyio==3.6.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
emoji==2.2.0
fonttools==4.38.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==2.6
isOdd==0.1.2
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
progress==1.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.0
requests==2.18.4
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
telebot==0.0.4
telebot7==0.1
urllib3==1.22
